hdf5ify/pyproject.toml
[tool.poetry]
name = "hdf5ify"
version = "0.1.0"
description = ""
authors = ["kerowak <boiwaq@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
h5py = "^3.6.0"
Pillow = "^8.4.0"
numpy = "^1.21.5"
common = { file = "../common/dist/common-0.1.0.tar.gz" }

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
